% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg-peaks.R
\name{.find_local_maxima}
\alias{.find_local_maxima}
\title{Find local maxima in a signal with minimum separation}
\usage{
.find_local_maxima(sig, min_distance)
}
\description{
Find local maxima in a signal with minimum separation
}
\keyword{internal}
