% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg-morphology.R
\name{.find_wave_peak}
\alias{.find_wave_peak}
\title{Find a wave peak (P or T) in a time window relative to R-peak}
\usage{
.find_wave_peak(
  sig,
  r_pos,
  win_start_samples,
  win_end_samples,
  n_time,
  find_max = TRUE
)
}
\description{
Find a wave peak (P or T) in a time window relative to R-peak
}
\keyword{internal}
