% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg-hrv-freq.R
\name{.hrv_lomb}
\alias{.hrv_lomb}
\title{Lomb-Scargle periodogram for HRV PSD estimation}
\usage{
.hrv_lomb(t_sec, rr_ms)
}
\arguments{
\item{t_sec}{Time stamps in seconds.}

\item{rr_ms}{RR intervals in milliseconds.}
}
\value{
List with freqs and psd vectors.
}
\description{
Lomb-Scargle periodogram for HRV PSD estimation
}
\keyword{internal}
