% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg-hrv-freq.R
\name{.hrv_welch}
\alias{.hrv_welch}
\title{Welch's method for HRV PSD estimation}
\usage{
.hrv_welch(t_sec, rr_ms, detrend = FALSE, lambda = 500)
}
\arguments{
\item{t_sec}{Time stamps in seconds.}

\item{rr_ms}{RR intervals in milliseconds.}

\item{detrend}{Logical; apply smoothness-priors detrending after resampling.}

\item{lambda}{Smoothing parameter for smoothness-priors detrending.}
}
\value{
List with freqs and psd vectors.
}
\description{
Welch's method for HRV PSD estimation
}
\keyword{internal}
