% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{PhysioECG-package}
\alias{PhysioECG}
\alias{PhysioECG-package}
\title{PhysioECG: ECG Analysis Functions for PhysioExperiment Objects}
\description{
Provides electrocardiography (ECG) analysis functions for PhysioExperiment objects. Includes R-peak detection (adaptive Pan-Tompkins), RR interval computation, HRV time-domain metrics (SDNN, RMSSD, pNN50), HRV frequency-domain analysis (VLF, LF, HF power), nonlinear HRV (Poincare, Sample Entropy, DFA), ECG morphology analysis (QRS boundaries, QT/QTc intervals), and signal quality assessment with ectopic beat correction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/x-biosignal/PhysioECG}
  \item \url{https://x-biosignal.r-universe.dev/PhysioECG}
  \item \url{https://x-biosignal.github.io/PhysioECG/}
  \item Report bugs at \url{https://github.com/x-biosignal/PhysioECG/issues}
}

}
\author{
\strong{Maintainer}: Yusuke Matsui \email{mail.to.matsui@gmail.com}

}
\keyword{internal}
